requirements.txt
nose==1.3.7
python-dateutil==2.4.2
requests==2.7.0
six==1.9.0
wheel==0.24.0
